Fixup the last test
This required moving the function to plugin.go. This makes sense,
because right now it is the only caller. Additionally adding it to the
API is probably not _that_ big a deal, but would have to be done first
because of the nature of the way the plugin loader work.